Split root Pipfile per test suite for true CI path-filter isolation
The path filters added in aecc573 still listed the shared root Pipfile/
Pipfile.lock (and best_practice/conftest.py) in every workflow, so any
dependency bump or shared-fixture edit still triggered every workflow -
not the isolation intended. Fixes it by giving each suite its own
Pipfile/Pipfile.lock (direct deps only, built from actual imports) and,
where fixtures were shared, its own conftest.py:

- best_practice/{desktop_web,mobile_native,mobile_web}
- examples/{selenium,playwright}
- examples/robotframework/{desktop_web,native_mobile}

Workflows now use per-suite working-directory and paths scoped to just
their own directory. python.yml is removed (its best-practice-desktop-us
leg was already redundant with desktop_web.yml; its demo leg moves to
the new selenium_demo.yml). Root Pipfile/Pipfile.lock are removed;
requirements.txt/.gitpod.yml are untouched (separate, unrelated
mechanism). sauceclient/sa11y are dropped as unused by any suite.
