+++ .github/workflows/desktop_web.yml
@@ -6,18 +6,12 @@ on:
         - main
     paths:
       - 'best_practice/desktop_web/**'
-      - 'best_practice/conftest.py'
-      - 'Pipfile'
-      - 'Pipfile.lock'
       - '.github/workflows/desktop_web.yml'
   pull_request:
     branches:
         - main
     paths:
       - 'best_practice/desktop_web/**'
-      - 'best_practice/conftest.py'
-      - 'Pipfile'
-      - 'Pipfile.lock'
       - '.github/workflows/desktop_web.yml'
   schedule:
     - cron: "0 9 * * *"
@@ -26,7 +20,10 @@ jobs:
   python-tests:
     name: Tests
     runs-on: ubuntu-latest
-    
+    defaults:
+      run:
+        working-directory: best_practice/desktop_web
+
     steps:
       - uses: actions/checkout@v7
       - name: Setup Python
