--- Pipfile
@@ -1,57 +1,0 @@
-[packages]
-pytest = "==9.1.1"
-pytest-xdist = "==3.8.0"
-Appium-Python-Client = "==5.3.1"
-PyYAML = "==6.0.3"
-selenium = "==4.46.0"
-sa11y = "*"
-sauceclient = "==1.0.2"
-robotframework = "==7.4.2"
-robotframework-pabot = "==5.2.2"
-robotframework-seleniumlibrary = "==6.9.0"
-apipkg = "==3.0.2"
-attrs = "==26.1.0"
-certifi = "==2026.7.22"
-chardet = "==7.4.3"
-decorator = "==5.3.1"
-docutils = "==0.23"
-execnet = "==2.1.2"
-idna = "==3.18"
-importlib-metadata = "==9.0.0"
-iniconfig = "==2.3.0"
-kitchen = "==1.2.6"
-more-itertools = "==11.1.0"
-packaging = "==26.2"
-pluggy = "==1.6.0"
-py = "==1.11.0"
-pyparsing = "==3.3.2"
-pytest-playwright = "==0.8.0"
-requests = "==2.34.2"
-robotframework-pythonlibcore = "==4.6.0"
-six = "==1.17.0"
-toml = "==0.10.2"
-urllib3 = "==2.7.0"
-zipp = "==4.1.0"
-playwright = "==1.61.0"
-
-[scripts]
-best-practice-desktop-us = "pytest -n8 --dc=us best_practice/desktop_web/"
-best-practice-desktop-eu = "pytest -n8 --dc=eu best_practice/desktop_web/"
-best-practice-mobile-web-vdc-us = "pytest -n8 --dc=us best_practice/mobile_web/vdc"
-best-practice-mobile-web-vdc-eu = "pytest -n8 --dc=eu best_practice/mobile_web/vdc"
-best-practice-mobile-web-rdc-us = "pytest -n8 --dc=us best_practice/mobile_web/rdc"
-best-practice-mobile-web-rdc-eu = "pytest -n8 --dc=eu best_practice/mobile_web/rdc"
-best-practice-mobile-native-us-android = "pytest -n8 --dc=us best_practice/mobile_native/android"
-best-practice-mobile-native-eu-android = "pytest -n8 --dc=eu best_practice/mobile_native/android"
-best-practice-mobile-native-us-ios = "pytest -n8 --dc=us best_practice/mobile_native/ios"
-best-practice-mobile-native-eu-ios = "pytest -n8 --dc=eu best_practice/mobile_native/ios"
-demo = "pytest examples/selenium"
-robot-desktop-web-us = "pabot --argumentfile1 examples/robotframework/desktop_web/chrome_config.txt --argumentfile2 examples/robotframework/desktop_web/firefox_config.txt --variable DATA_CENTER:'https://ondemand.us-west-1.saucelabs.com/wd/hub' examples/robotframework/desktop_web/Tests/"
-robot-desktop-web-eu = "pabot --argumentfile1 examples/robotframework/desktop_web/chrome_config.txt --argumentfile2 examples/robotframework/desktop_web/firefox_config.txt --variable DATA_CENTER:'https://ondemand.eu-central-1.saucelabs.com/wd/hub' examples/robotframework/desktop_web/Tests/"
-robot-android-us = "pabot --variable DATA_CENTER:'https://ondemand.us-west-1.saucelabs.com/wd/hub' -A examples/robotframework/native_mobile/android/android_config.txt examples/robotframework/native_mobile/android/tests/"
-robot-android-eu = "pabot --variable DATA_CENTER:'https://ondemand.eu-central-1.saucelabs.com/wd/hub' -A examples/robotframework/native_mobile/android/android_config.txt examples/robotframework/native_mobile/android/tests/"
-robot-ios-us = "pabot --variable DATA_CENTER:'https://ondemand.us-west-1.saucelabs.com/wd/hub' -A examples/robotframework/native_mobile/ios/iphone_config.txt examples/robotframework/native_mobile/ios/tests/"
-robot-ios-eu = "pabot --variable DATA_CENTER:'https://ondemand.eu-central-1.saucelabs.com/wd/hub' -A examples/robotframework/native_mobile/ios/iphone_config.txt examples/robotframework/native_mobile/ios/tests/"
-playwright-tests = "pytest -n8 examples/playwright/"
-
-[dev-packages]
