+++ best_practice/README.md
@@ -2,22 +2,26 @@
 
 This section demonstrates best practice examples of how to connect and run simple tests using Python on Sauce Labs. As a best practice, we recommend [Pytest](http://pytest.org) as a test framework and runner.
 
+Each suite below (`desktop_web`, `mobile_web`, `mobile_native`) has its own `Pipfile` and its
+own `conftest.py`, so dependencies stay scoped to just what that suite needs.
+
 ## Quickstart
 
 To run tests, do the following steps:
 
 1. Set your Sauce username and access key as environment variables on the machine you are running tests from as `SAUCE_USERNAME` and `SAUCE_ACCESS_KEY` respectively.
-2. Run 
+2. `cd` into the suite you want to run, e.g. `cd best_practice/desktop_web`.
+3. Run
 ```
 pipenv install
 ```
-3. Run 
+4. Run
 ```
-pipenv best-practice-desktop-us
-``` 
-to run against a desktop web app using Sauce Labs' US data center or 
+pipenv run best-practice-desktop-us
 ```
-pipenv best-practice-desktop-eu
+to run against a desktop web app using Sauce Labs' US data center or
+```
+pipenv run best-practice-desktop-eu
 ```
 to run against the EU data center
 
@@ -28,24 +32,28 @@ These tests use Pytest and are configured to run using [Pipenv](https://pipenv.p
 - Python 3.7+ (_NOTE_ these tests are only compatible with Python 3, and have only been tested against Python 3.7+)
 - Pipenv
 
-To install dependencies needed in this project, run 
+To install dependencies needed for a suite, run
 ```
 pipenv install
-``` 
-anywhere in this project. You can also install Pytest and Selenium directly if you do not want to use Pipenv.
+```
+from inside that suite's directory (`best_practice/desktop_web`, `best_practice/mobile_web`, or
+`best_practice/mobile_native`). You can also install Pytest and Selenium/Appium-Python-Client
+directly if you do not want to use Pipenv.
 
 Last and certainly not least, set your Sauce username and access key as environment variables on the machine you are running tests from as `SAUCE_USERNAME` and `SAUCE_ACCESS_KEY` respectively.
 
 ## Structure of the Best Practice tests
 
-These tests are designed to illustrate how to connect Pytest tests to running on Sauce Labs. 
+These tests are designed to illustrate how to connect Pytest tests to running on Sauce Labs.
 
-Pytest fixtures are located in the `conftest.py` file as is conventional with Pytest. The `conftest.py` file shows how to configure the Webdriver to connect to Sauce Labs via Capabilities.
+Each suite has its own `conftest.py`, as is conventional with Pytest. That `conftest.py` file
+shows how to configure the Webdriver to connect to Sauce Labs via Capabilities.
 
-Each directory contains tests that use fixtures from the `conftest.py` file. Each directory contains tests that test against
+Each directory contains tests that use fixtures from its own `conftest.py` file. Each directory
+contains tests that test against
 
 - desktop web browsers,
-- mobile web browsers (on emulators and simulators), and 
+- mobile web browsers (on emulators and simulators), and
 - a mobile native app on real devices.
 
 All tests make use of the Sauce Demo app, in both [web](https://github.com/saucelabs/sample-app-web) and [native mobile](https://github.com/saucelabs/sample-app-mobile) versions as the app under test.
@@ -58,84 +66,84 @@ Sauce currently supports multiple data centers, so you can choose where your tes
 
 ### US Case
 
-To run desktop web tests you can run 
+To run desktop web tests, from `best_practice/desktop_web`, run
 ```
-pipenv best-practice-desktop-us
-``` 
-or you can run 
+pipenv run best-practice-desktop-us
 ```
-pytest best-practice-us
-``` 
-directly if you have Pytest and Selenium installed. 
-
-Similarly you can run run 
+or run
 ```
-pipenv best-practice-mobile-web-us
-``` 
-or you can run 
-```
-pytest best-practice-mobile-us
+pytest -n8 --dc=us .
 ```
 directly if you have Pytest and Selenium installed.
 
+Similarly, from `best_practice/mobile_web`, run
+```
+pipenv run best-practice-mobile-web-vdc-us
+```
+or run
+```
+pytest -n8 --dc=us vdc
+```
+directly if you have Pytest and Selenium/Appium-Python-Client installed.
+
 To run native mobile app tests against real devices on the Sauce Real Device Cloud (RDC), confirm you have access to real devices on your account, upload the appropriate version of the [sample mobile app](https://github.com/saucelabs/sample-app-mobile/releases) to your account following [this guide](https://wiki.saucelabs.com/display/DOCS/Application+Storage#ApplicationStorage-WhatYou'llNeed).
 
-After this, you can run 
+After this, from `best_practice/mobile_native`, run
 ```
-pipenv best-practice-native-us-android
+pipenv run best-practice-mobile-native-us-android
 ```
-for the Android case (or 
+for the Android case (or
 ```
-pytest mobile_native/android
-``` 
-using Pytest directly) and/or 
+pytest -n8 --dc=us android
 ```
-pipenv best-practice-native-us-ios
-``` 
-for the iOS case (or 
+using Pytest directly) and/or
 ```
-pytest mobile_native/ios
-``` 
+pipenv run best-practice-mobile-native-us-ios
+```
+for the iOS case (or
+```
+pytest -n8 --dc=us ios
+```
 using Pytest directly)
 
 ### EU Case
 
-To run desktop web tests you can run 
+To run desktop web tests, from `best_practice/desktop_web`, run
 ```
-pipenv best-practice-desktop-eu
-``` 
-or you can run 
+pipenv run best-practice-desktop-eu
 ```
-pytest best-practice-eu
-``` 
-directly if you have Pytest and Selenium installed. 
-
-Similarly you can run run 
+or run
 ```
-pipenv best-practice-mobile-web-eu
-``` 
-or you can run 
+pytest -n8 --dc=eu .
 ```
-pytest best-practice-mobile-eu
-``` 
 directly if you have Pytest and Selenium installed.
+
+Similarly, from `best_practice/mobile_web`, run
+```
+pipenv run best-practice-mobile-web-vdc-eu
+```
+or run
+```
+pytest -n8 --dc=eu vdc
+```
+directly if you have Pytest and Selenium/Appium-Python-Client installed.
 
 To run native mobile app tests against real devices on the Sauce Real Device Cloud (RDC), confirm you have access to real devices on your account, upload the appropriate version of the [sample mobile app](https://github.com/saucelabs/sample-app-mobile/releases) to your account following [this guide](https://wiki.saucelabs.com/display/DOCS/Application+Storage#ApplicationStorage-WhatYou'llNeed).
 
-After this, you can run 
+After this, from `best_practice/mobile_native`, run
 ```
-pipenv best-practice-native-eu-android
-``` 
-for the Android case (or 
+pipenv run best-practice-mobile-native-eu-android
 ```
-pytest mobile_native/android
+for the Android case (or
 ```
-using Pytest directly) and/or 
+pytest -n8 --dc=eu android
 ```
-pipenv best-practice-native-eu-ios
+using Pytest directly) and/or
 ```
-for the iOS case (or 
+pipenv run best-practice-mobile-native-eu-ios
 ```
-pytest mobile_native/ios
+for the iOS case (or
+```
+pytest -n8 --dc=eu ios
 ```
 using Pytest directly)
