+++ best_practice/mobile_web/conftest.py
@@ -1,7 +1,6 @@
 import pytest
 from os import environ
 
-from appium.options.android import UiAutomator2Options
 from appium.options.common import AppiumOptions
 from appium.options.ios import XCUITestOptions
 from selenium import webdriver
@@ -40,37 +39,6 @@ emusim_browsers = [
     }]
 
 
-desktop_browsers = [
-    {
-        "platformName": "Windows 10",
-        "browserName": "MicrosoftEdge",
-        "platformVersion": "latest",
-        "sauce:options": {}
-    }, {
-        "platformName": "Windows 10",
-        "browserName": "firefox",
-        "platformVersion": "latest-1",
-        "sauce:options": {}
-    }, {
-        "platformName": "Windows 10",
-        "browserName": "internet explorer",
-        "platformVersion": "latest",
-        "sauce:options": {}
-    }, {
-        "platformName": "OS X 10.14",
-        "browserName": "safari",
-        "platformVersion": "latest-1",
-        "sauce:options": {}
-    }, {
-        "platformName": "OS X 10.14",
-        "browserName": "chrome",
-        "platformVersion": "latest",
-        "sauce:options": {
-            "extendedDebugging": True
-        }
-    }]
-
-
 def pytest_addoption(parser):
     parser.addoption("--dc", action="store", default='us', help="Set Sauce Labs Data Center (US or EU)")
 
@@ -85,10 +53,10 @@ def mobile_web_driver(request, data_center):
 
     test_name = request.node.name
     build_tag = environ.get('BUILD_TAG', "Sauce-Best-Practices-Python-Mobile-Web")
-   
+
     username = environ['SAUCE_USERNAME']
     access_key = environ['SAUCE_ACCESS_KEY']
-        
+
     if data_center and data_center.lower() == 'eu':
         selenium_endpoint = "https://@ondemand.eu-central-1.saucelabs.com/wd/hub"
     else:
@@ -133,46 +101,6 @@ def mobile_web_driver(request, data_center):
     browser.execute_script("sauce:job-result={}".format(sauce_result))
     browser.quit()
 
-@pytest.fixture(params=desktop_browsers)
-def desktop_web_driver(request, data_center):
-
-    test_name = request.node.name
-    build_tag = environ.get('BUILD_TAG', "Sauce-Best-Practices-Python-Desktop-Web")
-    
-    username = environ['SAUCE_USERNAME']
-    access_key = environ['SAUCE_ACCESS_KEY']
-    
-    if data_center and data_center.lower() == 'eu':
-        selenium_endpoint = "https://{}:{}@ondemand.eu-central-1.saucelabs.com/wd/hub".format(username, access_key)
-    else:
-        selenium_endpoint = "https://{}:{}@ondemand.us-west-1.saucelabs.com/wd/hub".format(username, access_key)
-
-    caps = dict()
-    caps.update(request.param)
-    caps['sauce:options'].update({'build': build_tag})
-    caps['sauce:options'].update({'name': test_name})
-
-    browser = webdriver.Remote(
-        command_executor=selenium_endpoint,
-        options=webdriver.ChromeOptions(),
-        keep_alive=True
-    )
-
-    # This is specifically for SauceLabs plugin.
-    # In case test fails after selenium session creation having this here will help track it down.
-    if browser is not None:
-        print("SauceOnDemandSessionID={} job-name={}".format(browser.session_id, test_name))
-    else:
-        raise WebDriverException("Never created!")
-
-    yield browser
-
-    # Teardown starts here
-    # report results
-    # use the test result to send the pass/fail status to Sauce Labs
-    sauce_result = "failed" if request.node.rep_call.failed else "passed"
-    browser.execute_script("sauce:job-result={}".format(sauce_result))
-    browser.quit()
 
 @pytest.fixture
 def rdc_browser(request, data_center):
@@ -203,66 +131,6 @@ def rdc_browser(request, data_center):
     driver.execute_script("sauce:job-result={}".format(sauce_result))
     driver.quit()
 
-@pytest.fixture
-def android_rdc_driver(request, data_center):
-
-    username_cap = environ['SAUCE_USERNAME']
-    access_key_cap = environ['SAUCE_ACCESS_KEY']
-
-    options = UiAutomator2Options()
-    options.platform_name = 'Android'
-    options.device_name = 'Google.*'
-    options.automation_name = 'UiAutomator2'
-    options.app = 'https://github.com/saucelabs/my-demo-app-android/releases/download/2.2.0/mda-2.2.0-25.apk'
-    sauce_options = {
-        'username': username_cap,
-        'accessKey': access_key_cap,
-        'build': 'RDC-Android-Python-Best-Practice',
-        'name': request.node.name,
-        'appiumVersion': 'latest',
-    }
-    options.set_capability('sauce:options', sauce_options)
-
-    if data_center and data_center.lower() == 'eu':
-        sauce_url = 'https://ondemand.eu-central-1.saucelabs.com/wd/hub'
-    else:
-        sauce_url = 'https://ondemand.us-west-1.saucelabs.com/wd/hub'
-
-    driver = appiumdriver.Remote(sauce_url, options=options)
-    yield driver
-    sauce_result = "failed" if request.node.rep_call.failed else "passed"
-    driver.execute_script("sauce:job-result={}".format(sauce_result))
-    driver.quit()
-
-@pytest.fixture
-def ios_rdc_driver(request, data_center):
-
-    username_cap = environ['SAUCE_USERNAME']
-    access_key_cap = environ['SAUCE_ACCESS_KEY']
-
-    options = XCUITestOptions()
-    options.platform_name = 'iOS'
-    options.device_name = 'iPhone.*'
-    options.app = 'https://github.com/saucelabs/sample-app-mobile/releases/download/2.7.1/iOS.RealDevice.SauceLabs.Mobile.Sample.app.2.7.1.ipa'
-    sauce_options = {
-        'username': username_cap,
-        'accessKey': access_key_cap,
-        'build': 'RDC-iOS-Python-Best-Practice',
-        'name': request.node.name,
-        'appiumVersion': 'latest',
-    }
-    options.set_capability('sauce:options', sauce_options)
-
-    if data_center and data_center.lower() == 'eu':
-        sauce_url = "https://ondemand.eu-central-1.saucelabs.com/wd/hub"
-    else:   
-        sauce_url = "https://ondemand.us-west-1.saucelabs.com/wd/hub"
-
-    driver = appiumdriver.Remote(sauce_url, options=options)
-    yield driver
-    sauce_result = "failed" if request.node.rep_call.failed else "passed"
-    driver.execute_script("sauce:job-result={}".format(sauce_result))
-    driver.quit()
 
 @pytest.hookimpl(hookwrapper=True, tryfirst=True)
 def pytest_runtest_makereport(item, call):
@@ -274,4 +142,3 @@ def pytest_runtest_makereport(item, call):
     # set an report attribute for each phase of a call, which can
     # be "setup", "call", "teardown"
     setattr(item, "rep_" + rep.when, rep)
-
