+++ examples/robotframework/native_mobile/Pipfile
@@ -1,0 +1,11 @@
+[packages]
+robotframework = "==7.4.2"
+robotframework-pabot = "==5.2.2"
+
+[scripts]
+robot-android-us = "pabot --variable DATA_CENTER:'https://ondemand.us-west-1.saucelabs.com/wd/hub' -A android/android_config.txt android/tests/"
+robot-android-eu = "pabot --variable DATA_CENTER:'https://ondemand.eu-central-1.saucelabs.com/wd/hub' -A android/android_config.txt android/tests/"
+robot-ios-us = "pabot --variable DATA_CENTER:'https://ondemand.us-west-1.saucelabs.com/wd/hub' -A ios/iphone_config.txt ios/tests/"
+robot-ios-eu = "pabot --variable DATA_CENTER:'https://ondemand.eu-central-1.saucelabs.com/wd/hub' -A ios/iphone_config.txt ios/tests/"
+
+[dev-packages]
